+++ ffgenerator.h
@@ -39,8 +39,8 @@ class FFGenerator {// Frittfall Generator
 		bool& boost() {return boost_;}
 		double& loss() { return loss_;}
 		void set_starthd(double y) { sy0_ = y;}
-		void set_aksellerasjon(float, float);
-		void set_aksellerasjon(float h, int a);
+		void set_aksellerasjon(double, double);
+		void set_aksellerasjon(double h, int a);
 	private:
 		//std::ofstream of {"dbg.txt"};
 //		double retning_nedover();
@@ -71,12 +71,12 @@ class FFGenerator {// Frittfall Generator
 	
 };
 
-void FFGenerator::set_aksellerasjon(float x, float y) {
+void FFGenerator::set_aksellerasjon(double x, double y) {
 	aksellerasjon_.X(x);
 	aksellerasjon_.Y(y);
 	
 }
-void FFGenerator::set_aksellerasjon(float h, int a) {
+void FFGenerator::set_aksellerasjon(double h, int a) {
 	aksellerasjon_.angle(a);
 	aksellerasjon_.hyp(h);
 	aksellerasjon_.calc_XY();
